requirements.txt
behave
websockets
confluent-kafka
pyhamcrest